Crate buck2_health_check_proto

Modules

health_check_client

Generated client implementations.

health_check_context_event

Nested message and enum types in HealthCheckContextEvent.

health_check_server

Generated server implementations.

remediation

Nested message and enum types in Remediation.

Structs

DisplayReport

A report that can be displayed to the user, typically on the console.

Empty

HealthCheckContextEvent

Event to update the HealthCheckContext in the server. This is a subset of the data from buck2 daemon events to avoid gRPC overheads.

HealthCheckResult

HealthIssue

Remediation

/ Remediation for an issue reported by a health check.

Report

Enums

HealthCheckType

Type of health check.

Severity

Severity of an issue reported by a health check.